# apps/api-gateway/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/accounts/{id}": {
      "get": {
        "operationId": "getAccountById",
        "summary": "Get an user account by account id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountDto"
                }
              }
            }
          }
        },
        "tags": [
          "Accounts"
        ]
      }
    },
...
    "/api/fields/create": {
      "post": {
        "operationId": "createField",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFieldBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Fields"
        ]
      }
    },
    "/api/fields/{id}": {
      "get": {
        "operationId": "getFieldById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Fields"
        ]
      }
    },
    "/api/fields/update/{id}": {
      "put": {
        "operationId": "updateFieldById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFieldBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Fields"
        ]
      }
    },
    "/api/fields/bulk/delete": {
      "post": {
        "operationId": "bulkDeleteFieldById",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteFieldsByIdBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Fields"
        ]
      }
    },
    "/api/fields/delete/{id}": {
      "delete": {
        "operationId": "deleteFieldById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Fields"
        ]
      }
    },
...
    "/api/invitations/create": {
      "post": {
        "operationId": "createInvitation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Invitations"
        ]
      }
    },
    "/api/invitations/accept/{id}": {
      "put": {
        "operationId": "acceptInvitationById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Invitations"
        ]
      }
    },
    "/api/invitations/bulk/reject": {
      "put": {
        "operationId": "bulkRejectInvitationsById",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkRejectInvitationsByIdBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Invitations"
        ]
      }
    },
    "/api/invitations/reject/{id}": {
      "put": {
        "operationId": "rejectInvitationById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Invitations"
        ]
      }
    },
...
    "/api/profiles/create": {
      "post": {
        "operationId": "registerProfile",
        "summary": "Register a new account's profile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProfileBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Registered",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/profiles/{id}": {
      "get": {
        "operationId": "getProfileById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/profiles/accounts/{accountId}": {
      "get": {
        "operationId": "getProfileByAccountId",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/profiles/update/{id}": {
      "put": {
        "operationId": "updateProfileById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
    "/api/profiles/update/accounts/{accountId}": {
      "put": {
        "operationId": "updateProfileByAccountId",
        "summary": "Update an account's profile by account id",
        "parameters": [
          {
            "name": "accountId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileDto"
                }
              }
            }
          }
        },
        "tags": [
          "Profiles"
        ]
      }
    },
...
    "/api/specializations/create": {
      "post": {
        "operationId": "createSpecialization",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSpecializationBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Specializations"
        ]
      }
    },
    "/api/specializations/{id}": {
      "get": {
        "operationId": "getSpecializationById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Specializations"
        ]
      }
    },
    "/api/specializations/update/{id}": {
      "put": {
        "operationId": "updateSpecializationById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSpecializationBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Specializations"
        ]
      }
    },
    "/api/specializations/bulk/delete": {
      "post": {
        "operationId": "bulkDeleteSpecializationById",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteSpecializationsByIdBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Specializations"
        ]
      }
    },
    "/api/specializations/delete/{id}": {
      "delete": {
        "operationId": "deleteSpecializationById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Specializations"
        ]
      }
    },
...
  },
  "info": {
    "title": "SV Connect API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
      "AccountDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "type": "object"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "email",
          "emailVerified",
          "role"
        ]
      },
...
      "ConnectSpecializationBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
      "CreateFieldBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "specializations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectSpecializationBody"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateFieldBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "specializations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectSpecializationBody"
            }
          }
        }
      },
      "BulkDeleteFieldsByIdBody": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "array"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
      "ConnectStudentBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "ConnectSupervisorBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        },
        "required": [
          "id"
        ]
      },
      "CreateInvitationBody": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED"
            ]
          },
          "student": {
            "$ref": "#/components/schemas/ConnectStudentBody"
          },
          "supervisor": {
            "$ref": "#/components/schemas/ConnectSupervisorBody"
          }
        },
        "required": [
          "student",
          "supervisor"
        ]
      },
      "BulkRejectInvitationsByIdBody": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
...
      "ConnectAccountBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
      "CreateProfileBody": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "headline": {
            "type": "object"
          },
          "summary": {
            "type": "object"
          },
          "pictureUrl": {
            "type": "object"
          },
          "backgroundUrl": {
            "type": "object"
          },
          "account": {
            "$ref": "#/components/schemas/ConnectAccountBody"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "account"
        ]
      },
      "ProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "headline": {
            "type": "object"
          },
          "summary": {
            "type": "object"
          },
          "pictureUrl": {
            "type": "object"
          },
          "backgroundUrl": {
            "type": "object"
          },
          "accountId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "firstName",
          "lastName",
          "headline",
          "summary",
          "pictureUrl",
          "backgroundUrl",
          "accountId"
        ]
      },
      "UpdateProfileBody": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "headline": {
            "type": "object"
          },
          "summary": {
            "type": "object"
          },
          "pictureUrl": {
            "type": "object"
          },
          "backgroundUrl": {
            "type": "object"
          }
        }
      },
      "ConnectFieldBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "CreateSpecializationBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectFieldBody"
            }
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateSpecializationBody": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConnectFieldBody"
            }
          }
        }
      },
      "BulkDeleteSpecializationsByIdBody": {
        "type": "object",
        "properties": {
          "ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "ids"
        ]
      },
...
    }
  }
}